requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
replicate==0.22.0
Pillow==10.1.0
requests==2.31.0
python-multipart==0.0.6
